views 까지 완성됨
