+++ wastory/api.py
@@ -8,7 +8,7 @@ from wastory.app.comment.views import comment_router
 from wastory.app.notification.views import notification_router
 from wastory.app.subscription.views import subscription_router
 from wastory.app.like.views import like_router
-
+from wastory.app.hometopic.views import hometopic_router
 
 api_router = APIRouter()
 
@@ -20,4 +20,4 @@ api_router.include_router(comment_router, prefix="/comments", tags=["comments"])
 api_router.include_router(notification_router, prefix="/notifications", tags=["notifications"])
 api_router.include_router(subscription_router, prefix="/subscription", tags=["subscriptions"])
 api_router.include_router(like_router, prefix="/likes", tags=["likes"])
-
+api_router.include_router(hometopic_router,prefix="/hometopics",tags=["hometopcs"])
